refactor: moved services to service folder
