+++ discord-bot/services/command-service.ts
