+++ discord-bot/services/event-service.ts
